miner/Miner-main/requirements.txt
GitPython==3.1.43
psutil==5.9.8
